backend/requirements.txt
fastapi
uvicorn[standard]
azure-data-tables
pydantic
python-jose[cryptography]
bcrypt
